# specs/001-money-ocr-api/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Money OCR API
  description: |
    金额识别OCR服务API

    提供基于PP-OCRv4的图片金额识别能力,支持单张和批量识别。

    ## 特性
    - 支持JPEG、PNG、BMP、TIFF格式
    - 自动去除货币符号和千分位分隔符
    - 返回纯数字格式金额
    - 提供置信度和处理时间
    - 健康检查端点

    ## 限制
    - 单张图片最大10MB
    - 响应时间<3秒
    - 建议批量请求≤10张图片
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8000/api/v1
    description: 本地开发环境
  - url: https://ocr-api.example.com/api/v1
    description: 生产环境

tags:
  - name: recognition
    description: 金额识别相关接口
  - name: health
    description: 健康检查接口

paths:
  /recognize:
    post:
      tags:
        - recognition
      summary: 单张图片金额识别
      description: 识别单张图片中的金额,返回纯数字格式
      operationId: recognizeAmount
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: 图片文件(JPEG/PNG/BMP/TIFF,最大10MB)
            encoding:
              file:
                contentType: image/jpeg, image/png, image/bmp, image/tiff
      responses:
        '200':
          description: 识别成功(包括未识别到金额的情况)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecognitionResponse'
              examples:
                success:
                  summary: 识别成功
                  value:
                    success: true
                    data:
                      amount: "1234.56"
                      confidence: 0.95
                      processing_time_ms: 1234
                      raw_text: "¥1,234.56"
                      warnings: []
                no_amount:
                  summary: 未识别到金额
                  value:
                    success: true
                    data:
                      amount: null
                      confidence: 0.0
                      processing_time_ms: 892
                      raw_text: ""
                      warnings: ["未识别到金额"]
                low_confidence:
                  summary: 置信度较低
                  value:
                    success: true
                    data:
                      amount: "5000"
                      confidence: 0.72
                      processing_time_ms: 2156
                      raw_text: "5000"
                      warnings: ["置信度较低,建议人工复核"]
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                no_file:
                  summary: 未提供文件
                  value:
                    success: false
                    error:
                      code: "NO_FILE_PROVIDED"
                      message: "请上传图片文件"
                unsupported_format:
                  summary: 不支持的格式
                  value:
                    success: false
                    error:
                      code: "UNSUPPORTED_FORMAT"
                      message: "仅支持JPEG、PNG、BMP、TIFF格式"
                      details: "当前格式: image/gif"
                invalid_image:
                  summary: 图片文件损坏
                  value:
                    success: false
                    error:
                      code: "INVALID_IMAGE"
                      message: "无法解析图片文件"
                      details: "图片头部数据损坏"
        '413':
          description: 文件过大
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "FILE_TOO_LARGE"
                  message: "图片大小超过10MB限制"
                  details: "当前大小: 12.5MB, 限制: 10MB"
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "OCR_ENGINE_ERROR"
                  message: "OCR处理失败"
                  details: "PaddleOCR引擎异常"
        '504':
          description: 请求超时
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "TIMEOUT"
                  message: "图片处理超时(>3秒)"
                  details: "处理时间: 3542ms"

  /recognize/batch:
    post:
      tags:
        - recognition
      summary: 批量图片金额识别
      description: 一次识别多张图片中的金额,返回每张图片的结果
      operationId: recognizeBatch
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - files
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: 图片文件列表(每张最大10MB,建议≤10张)
            encoding:
              files:
                contentType: image/jpeg, image/png, image/bmp, image/tiff
      responses:
        '200':
          description: 批量识别完成(部分成功也返回200)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchRecognitionResponse'
              example:
                success: true
                data:
                  total: 3
                  succeeded: 2
                  failed: 1
                  results:
                    - index: 0
                      filename: "invoice1.jpg"
                      success: true
                      data:
                        amount: "1234.56"
                        confidence: 0.95
                        processing_time_ms: 1234
                      error: null
                    - index: 1
                      filename: "invoice2.png"
                      success: false
                      data: null
                      error:
                        code: "FILE_TOO_LARGE"
                        message: "图片大小超过10MB限制"
                        details: "当前大小: 12.5MB"
                    - index: 2
                      filename: "receipt.jpg"
                      success: true
                      data:
                        amount: "5000"
                        confidence: 0.88
                        processing_time_ms: 2100
                      error: null
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "NO_FILE_PROVIDED"
                  message: "请至少上传一张图片"
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /health:
    get:
      tags:
        - health
      summary: 健康检查
      description: 返回服务运行状态和基本信息
      operationId: healthCheck
      responses:
        '200':
          description: 服务状态(健康或异常)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthCheckResponse'
              examples:
                healthy:
                  summary: 服务健康
                  value:
                    status: "healthy"
                    service: "money-ocr-api"
                    version: "1.0.0"
                    ocr_engine: "paddleocr-2.7.0"
                    uptime_seconds: 86400
                unhealthy:
                  summary: 服务异常
                  value:
                    status: "unhealthy"
                    service: "money-ocr-api"
                    version: "1.0.0"
                    ocr_engine: "paddleocr-2.7.0"
                    uptime_seconds: 3600
                    error: "OCR引擎初始化失败"

components:
  schemas:
    RecognitionResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          description: 请求是否成功
          example: true
        data:
          $ref: '#/components/schemas/RecognitionResult'

    RecognitionResult:
      type: object
      required:
        - amount
        - confidence
        - processing_time_ms
      properties:
        amount:
          type: string
          nullable: true
          description: 识别出的金额(纯数字格式),未识别到为null
          example: "1234.56"
          pattern: '^\d+(\.\d{1,2})?$'
        confidence:
          type: number
          format: float
          description: OCR识别置信度(0-1)
          example: 0.95
          minimum: 0.0
          maximum: 1.0
        processing_time_ms:
          type: integer
          description: 处理耗时(毫秒)
          example: 1234
          minimum: 0
        raw_text:
          type: string
          description: OCR原始识别文本(可选,用于调试)
          example: "¥1,234.56"
        warnings:
          type: array
          items:
            type: string
          description: 警告信息列表(可选)
          example: ["置信度较低,建议人工复核"]

    BatchRecognitionResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          description: 至少一个成功即为true
          example: true
        data:
          $ref: '#/components/schemas/BatchRecognitionResult'

    BatchRecognitionResult:
      type: object
      required:
        - total
        - succeeded
        - failed
        - results
      properties:
        total:
          type: integer
          description: 总图片数量
          example: 3
          minimum: 0
        succeeded:
          type: integer
          description: 成功识别数量
          example: 2
          minimum: 0
        failed:
          type: integer
          description: 失败数量
          example: 1
          minimum: 0
        results:
          type: array
          items:
            $ref: '#/components/schemas/BatchItemResult'
          description: 每张图片的识别结果

    BatchItemResult:
      type: object
      required:
        - index
        - filename
        - success
      properties:
        index:
          type: integer
          description: 图片索引(从0开始)
          example: 0
          minimum: 0
        filename:
          type: string
          description: 原始文件名
          example: "invoice1.jpg"
        success:
          type: boolean
          description: 该图片是否识别成功
          example: true
        data:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/RecognitionResult'
          description: 识别结果(成功时)
        error:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/ErrorDetail'
          description: 错误信息(失败时)

    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          description: 固定为false
          example: false
        error:
          $ref: '#/components/schemas/ErrorDetail'

    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: 错误码(大写下划线命名)
          example: "FILE_TOO_LARGE"
          enum:
            - FILE_TOO_LARGE
            - UNSUPPORTED_FORMAT
            - INVALID_IMAGE
            - NO_FILE_PROVIDED
            - OCR_ENGINE_ERROR
            - TIMEOUT
            - INTERNAL_ERROR
        message:
          type: string
          description: 人类可读的错误信息
          example: "图片大小超过10MB限制"
        details:
          type: string
          description: 详细错误信息(可选,用于调试)
          example: "当前大小: 12.5MB, 限制: 10MB"

    HealthCheckResponse:
      type: object
      required:
        - status
        - service
        - version
        - ocr_engine
        - uptime_seconds
      properties:
        status:
          type: string
          description: 服务状态
          example: "healthy"
          enum:
            - healthy
            - unhealthy
        service:
          type: string
          description: 服务名称
          example: "money-ocr-api"
        version:
          type: string
          description: 服务版本号
          example: "1.0.0"
        ocr_engine:
          type: string
          description: OCR引擎名称和版本
          example: "paddleocr-2.7.0"
        uptime_seconds:
          type: integer
          description: 服务运行时长(秒)
          example: 86400
          minimum: 0
        error:
          type: string
          description: 错误信息(异常时)
          example: "OCR引擎初始化失败"